requirements.txt
# Requirements for diffusion-policy-drake
# Generated from poetry env (Python 3.10)
#
# python3 -m venv env --without-pip
# source env/bin/activate
# curl -sS https://bootstrap.pypa.io/get-pip.py -o /tmp/get-pip.py
# python3 /tmp/get-pip.py --no-warn-script-location
# pip install -r requirements.txt
#
# Then install editable packages from their repos:
#   pip install -e .                                          # this repo (planning_through_contact)
#   pip install -e /data/locomotion/michzeng/diffusion-policy-experiments     # diffusion_policy
#   pip install -e /data/locomotion/michzeng/gcs-planar-pushing               # gcs_planar_pushing
#
# Drake must be Python 3.10 compatible; install via:
#   pip install drake==1.48.0

absl-py==2.3.1
accelerate==0.13.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.24.96
botocore==1.27.96
certifi==2025.11.12
cffi==1.15.1
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
Cython==0.29.32
dask==2024.1.0
datasets==2.6.1
debugpy==1.8.19
decorator==5.2.1
diffusers==0.11.1
dill==0.3.5.1
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.9
docker-pycreds==0.4.0
drake==1.48.0
einops==0.4.1
egl_probe==1.0.2
etils==1.13.0
exceptiongroup==1.3.1
fancycompleter==0.9.1
Farama-Notifications==0.0.4
fasteners==0.20
fastjsonschema==2.21.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.20.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
gradescope-utils==0.5.0
grpcio==1.76.0
gymnasium==1.2.3
h5py==3.7.0
huggingface-hub==0.25.2
hydra-core==1.2.0
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib_metadata==8.7.1
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.31.0
ipython==8.37.0
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
labmaze==1.0.6
llvmlite==0.42.0
lxml==5.4.0
lxml_html_clean==0.4.3
manipulation==2025.1.3
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.7.0
matplotlib-inline==0.2.1
more-itertools==10.8.0
Mosek==11.0.24
mpld3==0.5.12
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.4.0
multidict==6.7.0
multiprocess==0.70.13
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.59.1
numcodecs==0.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
parameterized==0.9.0
parso==0.8.5
partd==1.4.2
pathspec==0.12.1
pathtools==0.1.2
pdbpp==0.10.3
pexpect==4.9.0
pillow==12.0.0
portalocker==3.2.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==5.9.2
pyarrow==22.0.0
pycparser==2.23
pydot==1.4.2
pygame==2.6.1
pymunk==6.2.1
PyOpenGL==3.1.10
pyparsing==3.3.1
pytest==7.4.4
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2025.2
PyVirtualDisplay==3.0
PyWavelets==1.8.0
PyYAML==6.0.3
r3m @ https://github.com/facebookresearch/r3m/archive/b2334e726887fa0206962d7984c69c5fb09cceab.tar.gz
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
responses==0.18.0
robomimic==0.3.0
scikit-image==0.19.3
scikit-video==1.1.11
scipy==1.15.3
sentry-sdk==2.48.0
setproctitle==1.3.7
setuptools<70.0.0
Shapely==1.8.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2025.5.10
timeout-decorator==0.5.0
tk==0.1.0
torch==2.3.1
torchvision==0.18.1
tornado==6.5.4
tqdm==4.64.1
triton==2.3.1
typing_extensions==4.15.0
tzdata==2025.3
underactuated==2024.12.30
urllib3==1.26.20
wandb==0.13.3
wcwidth==0.2.14
Werkzeug==3.1.4
wmctrl==0.5
wrapt==2.0.1
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
zarr==2.12.0
